bls_data_sync/requirements.txt
boto3==1.34.69
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0 